+++ bittide-experiments/bittide-experiments.cabal.license
@@ -1,0 +1,3 @@
+SPDX-FileCopyrightText: 2024 Google LLC
+
+SPDX-License-Identifier: CC0-1.0
