+++ elastic-buffer-sim/elastic-buffer-sim.cabal
@@ -5,7 +5,7 @@ License:             Apache-2.0
 license-file:        LICENSE
 author:              QBayLogic B.V.
 maintainer:          devops@qbaylogic.com
-Copyright:           Copyright © 2022 Google LLC
+Copyright:           Copyright © 2022-2024 Google LLC
 
 common common-options
   default-extensions:
@@ -77,6 +77,7 @@ library
     array,
     aeson,
     bittide,
+    bittide-experiments,
     bytestring,
     cassava,
     clash-cores,
@@ -92,8 +93,7 @@ library
     Bittide.Domain
     Bittide.Plot
     Bittide.Simulate
-    Bittide.Topology
-    Bittide.Topology.Graph
+    Bittide.Simulate.Topology
   default-language: Haskell2010
 
 executable sim
@@ -101,15 +101,13 @@ executable sim
   main-is: exe/Main.hs
   build-depends:
     aeson,
-    array,
-    random,
+    bittide-experiments,
     bytestring,
-    containers,
-    directory,
-    happy-dot,
-    filepath,
     elastic-buffer-sim,
-    optparse-applicative,
+    directory,
+    filepath,
+    happy-dot,
+    optparse-applicative
   default-language: Haskell2010
   default-extensions: ImplicitPrelude
   -- enable rtsopts so we can setup memory limits
@@ -121,6 +119,7 @@ executable plot
   build-depends:
     array,
     bittide,
+    bittide-experiments,
     bittide-instances,
     bytestring,
     cassava,
