+++ elastic-buffer-sim/src/Bittide/Simulate.hs
@@ -4,7 +4,7 @@ Provides a rudimentary simulation of elastic buffers.
 
 -}
 
--- SPDX-FileCopyrightText: 2022 Google LLC
+-- SPDX-FileCopyrightText: 2022-2024 Google LLC
 --
 -- SPDX-License-Identifier: Apache-2.0
